requirements.txt
selenium>=4.18.0
google-api-python-client>=2.116.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
